refactor: extract checkbox-group and radio-group core styles
