--- packages/checkbox-group/src/vaadin-checkbox-group-styles.js
@@ -1,37 +1,0 @@
-/**
- * @license
- * Copyright (c) 2018 - 2025 Vaadin Ltd.
- * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
- */
-import { css } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-
-export const checkboxGroupStyles = css`
-  :host {
-    display: inline-flex;
-  }
-
-  :host::before {
-    content: '\\2003';
-    width: 0;
-    display: inline-block;
-  }
-
-  :host([hidden]) {
-    display: none !important;
-  }
-
-  .vaadin-group-field-container {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-  }
-
-  [part='group-field'] {
-    display: flex;
-    flex-wrap: wrap;
-  }
-
-  :host(:not([has-label])) [part='label'] {
-    display: none;
-  }
-`;
