+++ packages/checkbox-group/src/vaadin-checkbox-group.js
@@ -9,8 +9,8 @@ import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { checkboxGroupStyles } from './styles/vaadin-checkbox-group-core-styles.js';
 import { CheckboxGroupMixin } from './vaadin-checkbox-group-mixin.js';
-import { checkboxGroupStyles } from './vaadin-checkbox-group-styles.js';
 
 /**
  * `<vaadin-checkbox-group>` is a web component that allows the user to choose several items from a group of binary choices.
