+++ packages/field-base/src/styles/group-core-styles.js
@@ -3,9 +3,9 @@
  * Copyright (c) 2017 - 2025 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { css } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { css } from 'lit';
 
-export const radioGroupStyles = css`
+export const group = () => css`
   :host {
     display: inline-flex;
   }
