+++ packages/radio-group/src/styles/vaadin-radio-group-core-styles.js
@@ -1,0 +1,8 @@
+/**
+ * @license
+ * Copyright (c) 2017 - 2025 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+import { group } from '@vaadin/field-base/src/styles/group-core-styles.js';
+
+export const radioGroupStyles = group('radio');
